fix exporting lanes to only export the local snaps
