+++ e2e/harmony/harmony-import.e2e.ts
@@ -1,8 +1,5 @@
 import chai, { expect } from 'chai';
-import path from 'path';
-
 import { HARMONY_FEATURE } from '../../src/api/consumer/lib/feature-toggle';
-import { SchemaName } from '../../src/consumer/component/component-schema';
 import Helper from '../../src/e2e-helper/e2e-helper';
 
 chai.use(require('chai-fs'));
@@ -17,35 +14,21 @@ describe('import component on Harmony', function () {
   after(() => {
     helper.scopeHelper.destroy();
   });
-  describe('workspace with standard components', () => {
+  describe('tag, export, clean scope objects, tag and export', () => {
     before(() => {
       helper.scopeHelper.setNewLocalAndRemoteScopesHarmony();
-      helper.bitJsonc.setupDefault();
-      helper.fixtures.populateComponents();
+      helper.bitJsonc.disablePreview();
+      helper.fixtures.populateComponents(1);
       helper.command.tagAllComponents();
       helper.command.exportAllComponents();
-      helper.scopeHelper.reInitLocalScopeHarmony();
+      helper.git.mimicGitCloneLocalProjectHarmony();
       helper.scopeHelper.addRemoteScope();
-      helper.command.importComponent('comp1');
+      helper.command.importAllComponents();
+      helper.fixtures.populateComponents(1, undefined, ' v2');
+      helper.command.tagAllComponents();
     });
-    it('should import successfully with the schema prop', () => {
-      const comp1 = helper.command.catComponent(`${helper.scopes.remote}/comp1@latest`);
-      expect(comp1).to.have.property('schema');
-      expect(comp1.schema).to.equal(SchemaName.Harmony);
-    });
-    it('bit status should work and not show modified', () => {
-      const status = helper.command.statusJson();
-      expect(status.modifiedComponent).to.be.empty;
-    });
-    it('should create symlink from the workspace node_modules to the component node_modules after bit link', () => {
-      // before the link, the node_modules is created for the dependencies. as such, it happens
-      // at the same time this link is generated, which causes race condition. in a real world
-      // this won't happen because "bit install" will create the node_modules dir before the link
-      // starts.
-      helper.command.link();
-      expect(
-        path.join(helper.scopes.localPath, `node_modules/@${helper.scopes.remote}/comp1/node_modules`)
-      ).to.be.a.directory();
+    it('should export with no errors about missing artifacts (pkg file) from the first tag', () => {
+      expect(() => helper.command.exportAllComponents()).to.not.throw();
     });
   });
 });
